backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# PDF Processing
pdf2image==1.16.3
PyMuPDF==1.23.8
pillow==10.2.0

# OCR - Handwriting Recognition
paddlepaddle==2.6.2
paddleocr==2.7.3

# Table Detection
opencv-python-headless==4.9.0.80
numpy==1.26.3

# Report Generation
reportlab==4.0.9
jinja2==3.1.3

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic==2.5.3
